Initial Routes and Components
Let me know if we should change this, so we can all be on the same page before we initialize the components. Right now the app goes right into the list of items as the homepage, but we could add a home component if we want some other page upon opening.